+++ TestBS/Sensors.h
@@ -3,7 +3,7 @@
 
 // structure for light sensor
 struct sensor_encoder_t{
-  int16_t  angle;
+  int32_t  angle;
   int16_t  rpm;
 };
 
@@ -13,7 +13,7 @@ Sensors(void); // constructeur
 bool getTouch(int button); // on récupè-re la valeur du contacteur en booléen
 int readEncoder(void); //lance la lecture des codeurs
 int getSonar(uint8_t num); // valeur du sonar
-int getEncoder(int *right,int *left); //
+int getEncoder(int32_t *right,int32_t *left); //
 private :
 };
